% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_label}
\alias{create_label}
\title{Create proper label for tooltips and plots}
\usage{
create_label(x)
}
\arguments{
\item{x}{scalar string}
}
\value{
A character vector of re-maped labels.
}
\description{
Create proper label for tooltips and plots
}
\examples{
create_label("obese_rate")
create_label(c("obese_rate", "income"))
}
