% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\docType{data}
\name{cydict}
\alias{cydict}
\title{Dictionary of Country Names}
\format{
A data frame with 262 rows and 5 variables:
\describe{
\item{id}{The country ID in Altair's geojson template.}
\item{altair}{The country name in Altair (Python).}
\item{obesity}{The country name in the obesity dataset.}
\item{world_bank}{The country name in World Bank dataset.}
\item{iso3c}{The ISO 3-letter country code.}
}
}
\source{
\url{https://www.who.int/data/gho/data/indicators/indicator-details/GHO/prevalence-of-obesity-among-adults-bmi-=-30-(age-standardized-estimate)-(-)}
}
\usage{
cydict
}
\description{
A mapping of country IDs and names across datasets.
}
\keyword{datasets}
