% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_bar_plot}
\alias{make_bar_plot}
\title{Create a Barplot of Country Obesity Rankings}
\usage{
make_bar_plot(
  .region = NULL,
  .year = 2016,
  .income = NULL,
  .sex = NULL,
  n = 10
)
}
\arguments{
\item{.region}{The region input callback (character vector)}

\item{.year}{The year input callback (integer vector)}

\item{.income}{The income group callback (character vector)}

\item{.sex}{The sex group callback (scalar character)}

\item{n}{a scalar representing the number of countries to chart.}
}
\value{
A plotly object.
}
\description{
Create a Barplot of Country Obesity Rankings
}
\examples{
make_bar_plot()
}
