% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_choropleth_plot}
\alias{make_choropleth_plot}
\title{Create a Choropleth Map of Obesity Rates}
\usage{
make_choropleth_plot(.region = NULL, .year = 2016, .income = NULL, .sex = NULL)
}
\arguments{
\item{.region}{The region input callback (character vector)}

\item{.year}{The year input callback (integer vector)}

\item{.income}{The income group callback (character vector)}

\item{.sex}{The sex group callback (scalar character)}
}
\value{
A plotly object.
}
\description{
Create a Choropleth Map of Obesity Rates
}
\examples{
make_choropleth_plot()
}
