% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_rate_data}
\alias{make_rate_data}
\title{Generate a Rate Data for Plotting}
\usage{
make_rate_data(grp, fltr, vals = "obese")
}
\arguments{
\item{grp}{A character string of categorical grouping variable names.}

\item{fltr}{A named list of filters, where each list element name is the
name of the variable and the element value is a string of values whose
membership is to be evaluated (see examples).}

\item{vals}{A character string of continuous variables for which a rate will
be calculated relative to population. Defaults to "obese."}
}
\value{
A dataframe with rates calculated for each value in vals.
}
\description{
Generate a Rate Data for Plotting
}
\examples{
grp <- c("sex", "country")
vals <- c("obese", "unemployed")
fltr <- list(
  region = c("South Asia", "Europe & Central Asia"),
  year = c(2014, 2015, 2016)
)
make_rate_data(grp, fltr, vals)
}
