% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_scatter_plot}
\alias{make_scatter_plot}
\title{Create a Scatter Map of Obesity Rates vs. Other Variables}
\usage{
make_scatter_plot(
  .region = NULL,
  .year = NULL,
  .income = NULL,
  .sex = NULL,
  .regressor = "smoke",
  .grouper = "sex"
)
}
\arguments{
\item{.region}{The region input callback (character vector)}

\item{.year}{The year input callback (integer vector)}

\item{.income}{The income group callback (character vector)}

\item{.sex}{The sex group callback (scalar character)}

\item{.regressor}{The regressor to be used in the scatter plot (character vector)}

\item{.grouper}{The attribute to be used for grouping the data in the scatter plot (character vector)}
}
\value{
A plotly object.
}
\description{
Create a Scatter Map of Obesity Rates vs. Other Variables
}
