% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\docType{data}
\name{ob}
\alias{ob}
\title{The Obesity Data}
\format{
A data frame with 23,138 rows and 18 variables: \describe{
\item{country}{The country name.} \item{year}{The year of stratum.}
\item{sex}{The sex of individuals in the stratum.} \item{iso2c}{The ISO
2-letter country code.} \item{iso3c}{The ISO 3-letter country code.}
\item{region}{The geogrpahic region of the country.} \item{capital}{The
capital city of the country.} \item{longitude}{The longitude of the
capital city.} \item{latitude}{The latitude of the capital city.}
\item{income}{The name of the country income group.} \item{lending}{The
name of the IMF country lending category.} \item{lifexp}{The country life
expectancy for the stratum.} \item{pop}{The population count of the
stratum (row).} \item{primedu}{The number of individuals in the stratum
(row) who have completed basic primary education} \item{smoke}{The number
of individuals in the stratum (row) who smoke.} \item{unemployed}{The
number of individuals in the stratum (row) who are unemployed.}
\item{obese}{The number of individuals in the stratum (row) who are
obese.} \item{none}{A placeholder column with level "All" for dashboarding
purposes only.} }
}
\source{
\describe{
\item{WHO Obesity Data}{\url{https://www.who.int/data/gho/data/indicators/indicator-details/GHO/prevalence-of-obesity-among-adults-bmi-=-30-(age-standardized-estimate)-(-)}}
\item{World Bank Indicators}{\url{https://data.worldbank.org/indicator}}
}
}
\usage{
ob
}
\description{
A dataset containing the obesity dataset for the years 1960-2019. The
original dataset has been joined with World Bank indicator data. Data are
partitioned into mutually exclusive strata (rows) so that they may be
aggregated according to user needs.
}
\keyword{datasets}
