% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rate}
\alias{rate}
\title{Safely Calculate a Rate}
\usage{
rate(x, y)
}
\arguments{
\item{x}{A numeric variable acting as a numerator.}

\item{y}{A numeric variable acting as a denominator.}
}
\value{
A scalar rate.
}
\description{
Calculates sum(x) / sum(y) for values for non-missing values in (x, y).
}
\examples{
x <- c(1, 2, NA, 4)
y <- c(5, NA, 6, 7)
rate(x, y)
}
