% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{remap_sex}
\alias{remap_sex}
\title{Remap the Sex Input Variable}
\usage{
remap_sex(x = NULL)
}
\arguments{
\item{x}{A scalar string indicating the radio button selection.}
}
\value{
A character vector.
}
\description{
Remaps to c("Male", "Female") if user selects "Both"
}
\examples{
remap_sex("Male")
remap_sex("Both")
remap_sex()
}
