% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_bar_plot}
\alias{make_bar_plot}
\title{Create a Barplot of Country Obesity Rankings}
\usage{
make_bar_plot(
  .region = NULL,
  .year = 2016,
  .income = NULL,
  .sex = NULL,
  .ascending = TRUE,
  .n = 10
)
}
\arguments{
\item{.region}{The region input callback (character vector)}

\item{.year}{The year input callback (integer vector)}

\item{.income}{The income group callback (character vector)}

\item{.sex}{The sex group callback (scalar character)}

\item{.ascending}{a logical indicating the selection for top or bottom for
the country rankings. Also accepts a character. Defaults to TRUE.}

\item{.n}{a scalar representing the number of countries to chart.}
}
\value{
A plotly object.
}
\description{
Create a Barplot of Country Obesity Rankings
}
\examples{
make_bar_plot()
make_bar_plot(.n = 5)
make_bar_plot(.n = 5, .ascending = FALSE)
make_bar_plot(.n = 5, .ascending = "FALSE")
}
