% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_ts_plot}
\alias{make_ts_plot}
\title{Create a Time Series of Obesity Rates}
\usage{
make_ts_plot(
  .year = 2010,
  .sex = NULL,
  .highlight_country = "Canada",
  .year_range = list(1975, 2016)
)
}
\arguments{
\item{.year}{The year input callback (integer vector)}

\item{.sex}{The sex group callback (scalar character)}

\item{.highlight_country}{The countries we want to highlight (character
vector)}

\item{.year_range}{The year range input callback (integer vector)}
}
\value{
A plotly object.
}
\description{
Create a Time Series of Obesity Rates
}
